move config parsing from inference to distributions, workflow
